Add soft delete functionality to Customer model
Soft delete functionality has been introduced to the Customer model by implementing the SoftDeletes trait. This update allows for the restoration and permanent deletion of customers, with corresponding methods and routes added to the CustomerController. A new migration has also been created to add the soft delete column to the customers table.